Eval run detail: side-by-side, nested in the explorer, real error detail
- Extract the run detail view into EvalRunDetail.vue; the /evals/runs/[id]
  page is now a thin wrapper around it.
- Runs opened from an agent's Evals panel render NESTED inside the
  Knowledge Explorer (embedded mode, Back returns to the runs table in
  place) instead of navigating away to /evals.
- Expanded cases lay out side by side: conversation on the left, checks
  on the right, meta footer spanning below.
- Rules on a stopped/errored case no longer spin forever: a terminal
  result with no persisted verdict renders "not evaluated" instead of
  pending.
- Judge failures now carry the underlying error (quota, auth,
  connectivity...) instead of a bare "Judge evaluation failed", and
  errored agent blocks surface as red error lines in the conversation.
- Compare-baseline dropdown options include the run's start time —
  re-runs of the same case set were previously indistinguishable.
- Runs deep-linked from elsewhere honor ?back=<internal path>.
